Move form.ts to lib folder
